refactor(store): the source-neutral spine, enforced instead of documented (GDK-273)
`docs/ARCHITECTURE.md:79` has said for months: "The rule that matters:
`internal/store` does not import `internal/jira`." Constitution Article 6 is
*The Store Is Source-Neutral*, and `internal/adf` was split out specifically to
keep it that way. The rule was broken, and nothing noticed, because nothing
checked:

    $ go list -f '{{join .Deps "\n"}}' ./internal/store | grep midagedev
    …/internal/atlhttp
    …/internal/fields
    …/internal/jira

`internal/store/fields.go` imports `internal/fields` for three source-neutral
helpers (`SpecIDs`, `Coalesce`, `IsFilledAny`). But `classify.go`, `discover.go`
and `editable.go` in that package import `internal/jira`, and a package-level
import is an import — so the SQLite spine carried Jira REST types and the
Atlassian HTTP transport on its graph. `net/http` came along with them, breaking
`ARCHITECTURE.md:73` as well.

The split is by symbol, not by file, because `editable.go` mixed both kinds:
`EditKind` and `ResolveEditableID` take Jira types, while `EditableAlias`,
`EditableAliases`, `FieldValue` and `ValueFromIDs` do not. `internal/fields`
keeps the neutral half; `internal/jirafields` takes the Jira-shaped half and its
tests. `coalesce.go` already explained that `SpecIDs` exists so store can
coalesce without the full Jira shape — the code intended this boundary and the
package never caught up.

No behaviour, signature or exported name changed: the non-test exported surface
diffs empty against HEAD. Sync, server and CLI import both packages, which is
what they should do; only store is constrained.

`tools/check-store-deps.sh` is the part that makes this stay fixed, registered on
the CI Go job and deliberately **not** path-scoped, since any edit anywhere in
the import graph can break it. Its failure message quotes ARCHITECTURE.md:79 so
the next person learns the rule from the red instead of guessing. It also refuses
`internal/atlhttp` and `net/http`, and its match arms are written so
`internal/jira` does not match `internal/jirafields` — the sibling this very fix
created. `--graph` prints the graph for any package, so the diagnostic and the
gate are one command.

FAIL-first, run by the lead rather than taken on report: copied into the
unmodified tree at 25ad707 the gate exits 1 and names all three forbidden
packages; in this tree it exits 0.

Test count is unchanged at 1478 with packages 29 -> 30 (the new package), which
is the check that a file did not get lost in the move.
