+++ internal/jirafields/classify_test.go
@@ -1,4 +1,4 @@
-package fields
+package jirafields
 
 import (
 	"testing"
